requirements.txt
environs
flask
flask-marshmallow
flask-migrate
flask-sqlalchemy
marshmallow-sqlalchemy
marshmallow_enum
pymysql
passlib
nose